Fix the duplication in the overlay when grid toggl
When the grid was toggled, the overlay would actually
draw the main drawing frame instead of just redrawing
itself with transparent pixels.
This was due to the drawingcontroller passing the ref
to the same frame object when initialiwing both
renders.

By the way, the overlay should probably be treated a
bit differently as redrawing all its transprent pixels
in that case is useless.