Fix frame preview toggle when moving
The hiddenFrames list is not updated correctly when a frame is moved.
Unmoved frames are mapped incorrectly and result in "undefined" entries
in the list, and the resuling list is left unsorted. Make sure unmoved
entries still return a value when mapping and sort the list after frames
have moved.