PWA Support for Piskel!
Hey there! I really love using Piskel, and I thought the user experience could be even stronger if Piskel was also installable as a PWA directly from the browser.

These additions also include support for offline use, using a Service Worker. This is a key part of what makes PWAs so cool. You can also install this on desktop, thanks to modern browser releases, so Piskel will also be able to open in it's own app window just like a native app can!
Testing this out on Localhost seemed to work great for both `grunt play` and `grunt serve`, however I'm not sure what else is part of your build process for pushing to the main site. I imagine if it works locally with Grunt, than it shouldn't have an issue with the main site.

About two versions away from now, with version 99, Chromium browsers will be adding support for the Window Controls Overlay display mode, for PWAs, and this allows the web app to take up even more space of the window at once. You can also enable which areas should or shouldn't be draggable for the title bar.
I also added support for that feature with these additions, so that will also look very futuristic! Feel free to learn more about the upcoming API on this article by the Google team at web.dev: https://web.dev/window-controls-overlay/

Thanks for working on such a great app over the years! I'm happy I gotten to the point in programming where I can finally be able to contribute to it :)